split the nocode screen into separate views to simplify the logic
